Write data in a json file
