+++ json/authors.json
@@ -1,0 +1,1 @@
+[{"class":"Author","id":187,"first_name":"Masashi","last_name":"Kishimoto","items":["#<Game:0x00000164d3851680>","#<Game:0x00000164d3851680>"]}]
