+++ json/games.json
@@ -1,0 +1,1 @@
+[{"class":"Game","label":null,"author":{"class":"Author","id":187,"first_name":"Masashi","last_name":"Kishimoto","items":["#<Game:0x00000164d3851680>","#<Game:0x00000164d3851680>"]},"genre":null,"publish_date":"2023-09-04","id":270,"archived":false,"multiplayer":"y","last_played_at":"2024-01-29","title":"FIFA 22"}]
